requirements.txt
flask
scikit-learn
joblib
pandas
numpy
yfinance
ta
python-telegram-bot==13.15